feat: 아카이브 위치로 조회 API

+++ src/main/java/com/example/demo/archive/dto/ArchiveResponse.java
@@ -18,9 +18,9 @@ public class ArchiveResponse {
 
     private User author;
 
-    private BigDecimal positionX;
+    private Double positionX;
 
-    private BigDecimal positionY;
+    private Double positionY;
 
     private String address;
